+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3.24] - 2022-11-30
+### Added
+- Added RDP algorithm to `imgprocessing.py`
+- Added epsilon parameter to contour detection components
+
 ## [0.0.3.23] - 2022-11-28
 ### Added
 - Added `qrcode` library to `ddu_ias_research_env.yml`